requirements.txt
# Core
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Training & augmentation
numpy
pandas
opencv-python
Pillow
scikit-learn
editdistance

# Torch utilities
tqdm
matplotlib

# Optional but recommended
albumentations
